修复了新建图层时，由于符号和标签为空而报错的BUG；修复了重建导出时，遇到WFS图层报错的BUG；修复了图层超时时间过短导致复杂的WFS图层无法加载的BUG；图层列表右键菜单顶部显示图层类型

+++ MapBoard.Core/Util/LayerUtility.cs
@@ -27,18 +27,21 @@ namespace MapBoard.Util
                 layers.Remove(layer);
             }
 
-            if (deleteFiles)
+            if (deleteFiles )
             {
-                await Task.Run(() =>
+                if (layer is ShapefileMapLayerInfo)
                 {
-                    foreach (var file in Shapefile.GetExistShapefiles(Parameters.DataPath, layer.Name))
+                    await Task.Run(() =>
                     {
-                        if (Path.GetFileNameWithoutExtension(file) == layer.Name)
+                        foreach (var file in Shapefile.GetExistShapefiles(Parameters.DataPath, layer.Name))
                         {
-                            File.Delete(file);
+                            if (Path.GetFileNameWithoutExtension(file) == layer.Name)
+                            {
+                                File.Delete(file);
+                            }
                         }
-                    }
-                });
+                    });
+                }
             }
         }
